cxx-qt-gen: add mutable as a field to ParsedQInvokable
This helps generation later as we can then simply read the bool
