+++ crates/cxx-qt-gen/src/generator/naming/invokable.rs
@@ -62,6 +62,7 @@ mod tests {
         });
         let parsed = ParsedQInvokable {
             method: item,
+            mutable: false,
             return_cxx_type: None,
         };
